Make record immutable to prevent any accidental bugs
It's better to just have an actual immutable record of the incoming state and use that to do all the calculations and set the new values on the actual items all the way at the end.
Otherwise, changes in the order of calculations (for example: updating sellIn before quality) would cause a bug.